Refactor GNSS IMU fusion into package modules
